pyproject.toml
[project]
name = "agnoai"
version = "0.1.0"
description = "AI agent and workflow creation using agno"
authors = [
    {name = "sponge-bobs-square-pants",email = "parthchawla65@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13 <4.0"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "yfinance (>=0.2.65,<0.3.0)",
    "duckduckgo-search (>=8.1.1,<9.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "groq (>=0.30.0,<0.31.0)",
    "openai (>=1.97.0,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "agno (>=1.7.5,<2.0.0)",
    "pymongo (>=4.13.2,<5.0.0)",
    "newspaper4k (>=0.9.3.1,<0.10.0.0)",
    "lxml-html-clean (>=0.4.2,<0.5.0)",
    "ddgs (>=9.3.1,<10.0.0)",
    "arxiv (>=2.2.0,<3.0.0)",
    "googlesearch-python (>=1.3.0,<2.0.0)",
    "pycountry (>=24.6.1,<25.0.0)",
    "pypdf (>=5.8.0,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
